requirements.txt
httpx>=0.28
matplotlib>=3.10
pandas>=2.3
psutil>=7.0
pyyaml>=6.0
